Move WzInfoButton
